Move waitlist DB logic to admin; proxy from evrydayarchive-web
evrydayarchive-web was directly importing @repo/db, which calls
getDbEnv() at module load time and throws if DATABASE_URL is not set.
The public site is not supposed to connect to the DB directly.

Moves the upsert + rate limiter to admin /api/public/waitlist and
rewrites the public-site route to proxy through ADMIN_API_BASE_URL.
